Split payment-processing.ts into a folder and remove its line-limit exemption
The 1,985-line payment-processing module was grandfathered out of the
noExcessiveLinesPerFile rule. Split it into a payment-processing/ folder of
single-purpose files, each well under the limit, and drop its biome.json entry:

- metadata.ts   — read signed session metadata into a BookingIntent
- classify.ts   — validate the paid session and prove it is ours (price proof)
- cancel.ts     — the payment-cancelled/declined page and its retry link
- refunds.ts    — refund mechanics plus the typed refund reasons (RefundSpec)
- package-pricing.ts — re-validate package structure and per-line price drift
- items.ts      — validate every signed line against the current database
- pricing.ts    — build the checkout intent, re-price, and gate on price drift
- create.ts     — turn a validated session into a real attendee
- store-refund.ts — keep-and-refund placeholder + balance settle paths
- index.ts      — the two-phase locked orchestration

Splitting surfaced duplication jscpd could not see inside the monolith; fixed
with shared helpers rather than ignores: bookingSlot (listing id + package
path), datelessGhostBookings, allocatedUnitsByChild, a SessionProcessor type
for the two processors, and merging currentOrderTree into orderEdgeDrifted.

Importers (webhooks.ts, the two test files) now import from the specific files.
