+++ biome.json
@@ -152,7 +152,6 @@
     },
     {
       "includes": [
-        "src/features/api/payment-processing.ts",
         "src/shared/db/migrations/schema.ts",
         "src/ui/templates/public/reservations.tsx",
         "test/lib/server-listing-parents.test.ts",
