+++ src/features/api/payment-processing/store-refund.ts
@@ -1,0 +1,220 @@
+/**
+ * The keep-and-refund paths of the payment machine. When a signed-by-us payment
+ * can't be honoured — a price changed, a listing was deleted, an extra sold out,
+ * the event filled, or an unexpected error hit after the charge — we never drop
+ * the customer: we store a quantity-0 placeholder, refund the payment, record the
+ * cash round-trip in the ledger, and flag the attendee with a plain-language
+ * note. A balance session settles the existing attendee instead of creating one.
+ */
+
+import {
+  attendeeBaseFields,
+  bookingSlot,
+  type HonourResult,
+  sessionSuccess,
+} from "#routes/api/payment-processing/create.ts";
+import { businessTime } from "#routes/api/payment-processing/metadata.ts";
+import type { ValidatedItem } from "#routes/api/payment-processing/package-pricing.ts";
+import {
+  capacitySpec,
+  type RefundSpec,
+  refundAndFail,
+  refundedNoteText,
+  soldOutSpec,
+  tryRefund,
+} from "#routes/api/payment-processing/refunds.ts";
+import type {
+  BookingIntent,
+  PaymentFailureResult,
+  PaymentResult,
+} from "#routes/api/webhook-types.ts";
+import { bookingDateFields } from "#routes/public/ticket-payment.ts";
+import { logActivity } from "#shared/db/activityLog.ts";
+import { settleAttendeeBalance } from "#shared/db/attendees/balance.ts";
+import { createAttendeeAtomic } from "#shared/db/attendees.ts";
+import { balanceFinalizeStatement } from "#shared/db/processed-payments.ts";
+import { createSystemNote } from "#shared/db/system-notes.ts";
+import { ErrorCode, logError } from "#shared/logger.ts";
+import { sendNtfyError } from "#shared/ntfy.ts";
+import type { BookingItem, ValidatedPaymentSession } from "#shared/payments.ts";
+import { addPendingWork } from "#shared/pending-work.ts";
+import { recordPlaceholderRefund } from "#shared/refund-ledger.ts";
+
+/** User-facing message when the outstanding balance changed mid-payment. */
+const BALANCE_CHANGED_MESSAGE =
+  "The outstanding balance for this booking changed while you were paying.";
+
+/**
+ * User-facing message when a signed-by-us payment can't be honoured (price
+ * changed, charge mismatch, sold out, or an unexpected error) so the booking is
+ * kept and refunded. The refund clause is appended by formatPaymentError (or the
+ * refund-pending suffix below), so this just covers "we saved your details".
+ */
+const BOOKING_SAVED_MESSAGE =
+  "We couldn't complete your booking, so we've saved your details and a member of our team can help you rebook.";
+
+/** The quantity-0, money-free booking lines for a stored-but-refunded placeholder
+ *  — one per validated item, carrying the listing's current date range so the
+ *  ghost still sits on the right day, and each line's package path so a listing
+ *  booked through two paths keeps two distinct slots (identical slots would be
+ *  refused as duplicates and crash the store-and-refund). */
+export const placeholderBookings = (
+  validatedItems: ValidatedItem[],
+  intent: BookingIntent,
+) =>
+  validatedItems.map(({ item, listing }) => ({
+    ...bookingSlot(item),
+    pricePaid: 0,
+    quantity: 0,
+    ...bookingDateFields(listing, intent.date, intent.dayCount),
+  }));
+
+/** Quantity-0 ghost rows for a since-deleted listing: no date fields, because
+ * the listing row is gone and there is nothing left to derive a range from. Used
+ * per SIGNED LINE so a multi-item cart's deleted line is named with its package
+ * path rather than collapsed onto the first listing. */
+export const datelessGhostBookings = (items: readonly BookingItem[]) =>
+  items.map((item) => ({ ...bookingSlot(item), pricePaid: 0, quantity: 0 }));
+
+type PlaceholderBookings = Parameters<
+  typeof createAttendeeAtomic
+>[0]["bookings"];
+
+/**
+ * Settle a reserved attendee's balance instead of creating a new attendee.
+ *
+ * Reached only for a trusted session (the mismatch verdict refunds upstream), so
+ * the proof has already bound `balance_attendee_id` and the single balance line,
+ * and the charge equals the signed total. The amount this checkout was created
+ * for is that line's price (`items[0].p`); the settle clears the balance only if
+ * the live `remaining_balance` still equals it — so a balance the owner edited,
+ * or one a concurrent/stale checkout already settled, can't be cleared for the
+ * wrong figure — and finalizes the session in the SAME transaction so a crash
+ * between settle and finalize can't leave a paid-but-unfinalized row (which a
+ * later stale-replay would wrongly refund). A mismatch refunds and returns a
+ * terminal failure rather than mutating anything.
+ */
+export const settleBalanceSession = async (
+  sessionId: string,
+  session: ValidatedPaymentSession,
+  intent: BookingIntent,
+): Promise<PaymentResult> => {
+  const attendeeId = intent.balanceAttendeeId as number;
+  // A balance checkout is always a single synthetic line whose price is the
+  // outstanding balance it was created to clear (proof-bound: see handleBalancePost).
+  const expectedAmount = intent.items[0]!.p;
+  const listingId = intent.items[0]!.e;
+
+  // settleAttendeeBalance posts the balance payment itself (world funds the
+  // attendee, zeroing what they owed) guarded on the ledger balance, keyed to
+  // this session so a webhook retry is a no-op. We only finalize the payment
+  // session here, atomically with the settle.
+  const settled = await settleAttendeeBalance(
+    attendeeId,
+    expectedAmount,
+    { id: sessionId, occurredAt: businessTime(session) },
+    [balanceFinalizeStatement(sessionId, attendeeId, expectedAmount)],
+  );
+  if (!settled.settled) {
+    return refundAndFail(
+      session,
+      BALANCE_CHANGED_MESSAGE,
+      listingId,
+      409,
+      `Balance not settled (${settled.reason}) for attendee ${attendeeId}; paid ${session.amountTotal}`,
+    );
+  }
+
+  // Settle + finalize already committed atomically above. The listing (which
+  // may since be deleted) is resolved lazily by the redirect for its thank-you
+  // link, so we carry only its id here.
+  return sessionSuccess(attendeeId, listingId);
+};
+
+/**
+ * Keep a signed-by-us booking we can't honour rather than dropping it into limbo:
+ * store it as a quantity-0 placeholder (overbook-tolerant, so capacity — or a
+ * since-deleted listing — can never downgrade the store into a drop), refund the
+ * payment, record the cash round-trip in the ledger (a `payment` + `refund_cash`
+ * with NO `sale` leg, so the placeholder recognises no revenue and its projected
+ * price_paid stays 0), and flag the attendee with a plain-language system note
+ * carrying the reason and the provider's payment reference. The customer is told
+ * their details were saved and the payment refunded; no ticket is issued.
+ *
+ * We never report `refunded: false`. The booking now exists, so a retry must NOT
+ * re-create it — an un-refunded payment is recorded as a terminal, operator-
+ * resolved outcome (the note's manual-refund instruction stands) rather than
+ * released for re-processing.
+ */
+export const storeRefundedBooking = async (
+  session: ValidatedPaymentSession,
+  intent: BookingIntent,
+  bookings: PlaceholderBookings,
+  spec: RefundSpec,
+): Promise<PaymentFailureResult> => {
+  if (spec.notify) addPendingWork(sendNtfyError(spec.notify));
+  const listingId = bookings[0]!.listingId;
+  // A quantity-0 overbook insert has no capacity gate and consumes no modifier
+  // stock, so it always writes the row — trust it. (If the PII can't encrypt the
+  // whole system is broken; we don't defend against that.)
+  const stored = await createAttendeeAtomic({
+    ...(await attendeeBaseFields(session, intent)),
+    allowOverbook: true,
+    bookings,
+  });
+  const attendeeId = (stored as Extract<typeof stored, { success: true }>)
+    .attendees[0]!.id;
+  const refunded = await tryRefund(session.paymentReference, listingId);
+  await recordPlaceholderRefund(
+    {
+      amount: session.amountTotal,
+      attendeeId,
+      eventId: session.id,
+      listingId,
+      occurredAt: businessTime(session),
+    },
+    spec.code,
+    refunded,
+  );
+  if (refunded) {
+    await logActivity(
+      `Automatic refund (${spec.code}); booking kept at quantity 0`,
+      listingId,
+      attendeeId,
+    );
+  } else {
+    logError({
+      code: ErrorCode.PAYMENT_REFUND,
+      detail: `Stored-but-unrefunded booking ${attendeeId} (${spec.code}): ${spec.detail}`,
+      listingId,
+    });
+  }
+  await createSystemNote(
+    attendeeId,
+    refundedNoteText(attendeeId, spec, refunded, session.paymentReference),
+  );
+  // Status 200: a fully-handled terminal outcome (booking kept, money returned or
+  // flagged). The webhook acks it (never the 409 transient-lock retry nor a 503
+  // refund retry — the booking exists, so a retry can't re-create it), and the
+  // customer sees an informational "saved your details" message.
+  return {
+    detail: spec.detail,
+    error: refunded
+      ? BOOKING_SAVED_MESSAGE
+      : `${BOOKING_SAVED_MESSAGE} Your refund is being arranged — please contact us if it does not arrive.`,
+    ...(refunded ? { refunded: true } : {}),
+    status: 200,
+    success: false,
+  };
+};
+
+/** The placeholder refund reason for a booking we tried but couldn't honour: a
+ *  sold-out extra reads differently from a full event, and anything else
+ *  (capacity, or the broken-system encryption_error we don't special-case) is
+ *  treated as "the event filled up". */
+export const specForFailure = (
+  failure: Extract<HonourResult, { ok: false }>,
+): RefundSpec =>
+  failure.reason === "sold_out"
+    ? soldOutSpec(failure.detail)
+    : capacitySpec(failure.detail);
